Add skeleton support for package subcommand.
We create a new archive format called DAR (deterministic archive) to package up
chaincode assets.  Our first approach was to reuse an existing archive format/tooling
(e.g. tar, zip, etc).  However, most had problems with determinisim.  For instance,
tar injects things like UIDs and timestamps into the binary, which could result
in a different binary each time regardless of code content.

The "ar" tool from binutils showed some promise with explicit support for determinism,
but its focus on elf containers seemed inappropriate.  In the end, we realized it
might be more straightforward to leverage our existing investment in protobufs and
build our own custom format.  It's not all that complicated anyway.

The current implementation is somewhat naive in that it assumes that the compressed
contents of the archive will have no problem fitting in memory and/or a single
protobuf object (defaults to 64MB max, I think).  This is ok.  For one, chaincode
packages by definition should strive to be as small as possible since they
need to be distributed around and persisted on the ledger.  If this ever changes,
the fundamental nature of protobufs should allow us to be creative in enhancing
the capabilities without breaking compatibility.

Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>