Explicitly set -server in the JVM options
It was discovered that chaintool would not run on Power
systems because the JVM would complain about the -client
switch being passed in.  This switch was emitted into
the JAR preamble script created by the lein-bin plugin.

It seems that the -client/-server option is a no-op on
modern 64 bit JVMs.  For some reason, the OpenJDK-8
on Power seems to reject any setting of -client whereas
this works fine on x86 and Z (even for the same exact
version of OpenJDK-8).  Therefore, we set the :jvm-opts
flag for the project which modifies how the lein-bin
plugin emits code to work around the issue.  If the
rumors regarding the efficacy of the switch are true,
this should have zero negative impact on Z/X86 while
allowing P to work around the problem.

Fixes Issue #28

Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>