Update to java 1.8
We had intentionally set compatibility with 1.7 previously
because the hyperledger fabric devenv included 1.7 by
default.  The main downside to this was that we were
missing out on the bootclasspath feature which is supposed
to help reduce start up times.  Now that more recent
versions of hyperledger include 1.8, its time to move
the ball forward.

Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>