Fix generation of client-stub.go for consumed chaincode -- fix "invoker"
The v0.9.x code generating template was producing calls in
client-stub.go using an incorrect signature for the following methods on
ChaincodeStubInterface (as of fabric v0.6.1-preview):

    InvokeChaincode(chaincodeName string, args [][]byte) ([]byte, error)
    QueryChaincode(chaincodeName string, args [][]byte) ([]byte, error)

These incorrect signatures were apparently part of an older API that
looked like the following:

    Invoke(chaincodeName string, function string, args []string) ([]byte, error)
    Query(chaincodeName string, function string, args []string) ([]byte, error)

As a secondary fix, the invoker example's Go source had the package name
"chaincode" instead of "main".  This made the `chaintool build` command
produce archive files instead of executable binaries as desired.  This
fix is included in this patch because it is needed to validate the
change made to the Go template.

Signed-off-by: Bryan Matsuo <bryan.matsuo@gmail.com>